requirements.txt
flask==3.1.1
flask-cors==5.0.1
pyTelegramBotAPI==4.26.0
Pillow==11.2.1
gunicorn==23.0.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
requests==2.32.3
